+++ backend/src/main/kotlin/com/book/manager/usecase/startRental/StartRentalUseCase.kt
@@ -1,28 +1,39 @@
-package com.book.manager.usecase.end.rental
+package com.book.manager.usecase.startRental
 
 import com.book.manager.domain.criteria.OperatorCriteria
 import com.book.manager.domain.model.id.BookId
 import com.book.manager.domain.model.id.OperatorId
+import com.book.manager.domain.model.value.Rental
 import com.book.manager.domain.repository.BookRepository
 import com.book.manager.domain.repository.OperatorRepository
 import com.book.manager.domain.repository.RentalRepository
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
+import java.time.LocalDateTime
 
 @Service
-class EndRentalUseCase(
+class StartRentalUseCase(
     private val operatorRepository: OperatorRepository,
     private val bookRepository: BookRepository,
     private val rentalRepository: RentalRepository
 ) {
+    companion object {
+        private const val RENTAL_TERM_DAYS = 14L
+    }
+
     @Transactional
     operator fun invoke(bookId: BookId, operatorId: OperatorId) {
         require(operatorRepository.find(OperatorCriteria(id = operatorId)) != null) { "Operator not found" }
         val book = bookRepository.findWithRental(bookId) ?: throw IllegalArgumentException("Book not found")
 
-        check(book.isRental) { "Book is not rented" }
-        check(book.rental!!.operatorId == operatorId) { "Operator is not the renter" }
+        require(!book.isRental) { "Book is already rented" }
 
-        rentalRepository.endRental(bookId)
+        val rentalDateTime = LocalDateTime.now()
+        val returnDeadline = rentalDateTime.plusDays(RENTAL_TERM_DAYS)
+        val rental = Rental(operatorId, rentalDateTime, returnDeadline)
+
+        val newBook = book.copy(rental = rental)
+
+        rentalRepository.startRental(newBook)
     }
 }
